// web/src/lib/api/generated/models/PaginatedJobs.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Service
 * The Scheduler manages periodic jobs and dispatches them to the worker queue when due. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Job } from './Job';
import {
    JobFromJSON,
    JobFromJSONTyped,
    JobToJSON,
    JobToJSONTyped,
} from './Job';

/**
 * 
 * @export
 * @interface PaginatedJobs
 */
export interface PaginatedJobs {
    /**
     * 
     * @type {Array<Job>}
     * @memberof PaginatedJobs
     */
    items: Array<Job>;
    /**
     * 
     * @type {number}
     * @memberof PaginatedJobs
     */
    page: number;
    /**
     * 
     * @type {number}
     * @memberof PaginatedJobs
     */
    pageSize: number;
    /**
     * 
     * @type {number}
     * @memberof PaginatedJobs
     */
    totalCount: number;
    /**
     * 
     * @type {number}
     * @memberof PaginatedJobs
     */
    totalPages: number;
}

/**
 * Check if a given object implements the PaginatedJobs interface.
 */
export function instanceOfPaginatedJobs(value: object): value is PaginatedJobs {
    if (!('items' in value) || value['items'] === undefined) return false;
    if (!('page' in value) || value['page'] === undefined) return false;
    if (!('pageSize' in value) || value['pageSize'] === undefined) return false;
    if (!('totalCount' in value) || value['totalCount'] === undefined) return false;
    if (!('totalPages' in value) || value['totalPages'] === undefined) return false;
    return true;
}

export function PaginatedJobsFromJSON(json: any): PaginatedJobs {
    return PaginatedJobsFromJSONTyped(json, false);
}

export function PaginatedJobsFromJSONTyped(json: any, ignoreDiscriminator: boolean): PaginatedJobs {
    if (json == null) {
        return json;
    }
    return {
        
        'items': ((json['items'] as Array<any>).map(JobFromJSON)),
        'page': json['page'],
        'pageSize': json['pageSize'],
        'totalCount': json['totalCount'],
        'totalPages': json['totalPages'],
    };
}

export function PaginatedJobsToJSON(json: any): PaginatedJobs {
    return PaginatedJobsToJSONTyped(json, false);
}

export function PaginatedJobsToJSONTyped(value?: PaginatedJobs | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'items': ((value['items'] as Array<any>).map(JobToJSON)),
        'page': value['page'],
        'pageSize': value['pageSize'],
        'totalCount': value['totalCount'],
        'totalPages': value['totalPages'],
    };
}